requirements.txt
streamlit
python-dotenv
seaborn
pandas
openai
git+https://github.com/openai/whisper.git
streamlit-audiorecorder
gradio 